+++ nyc_data/xlsx_utils.py
@@ -6,9 +6,10 @@ from typing import NamedTuple, Any, Callable, List, Optional, Set
 from django.core.serializers.json import DjangoJSONEncoder
 from openpyxl import load_workbook
 
+from ppe import errors
 from ppe.data_mapping.types import DataFile
 from ppe.data_mapping.utils import ErrorCollector
-from ppe import data_import
+
 
 def XLSXDictReader(sheet):
     rows = sheet.max_row
@@ -61,7 +62,7 @@ def guess_mapping(sheet: Path, possible_mappings: List[SheetMapping]):
                     first_row = next(reader)
                     
             except Exception as exc:
-                raise data_import.CsvImportError('Error reading in CSV file')
+                raise errors.CsvImportError('Error reading in CSV file') from exc
 
         col_names = [m.sheet_column_name for m in mapping.mappings]
         if all(col_name in first_row for col_name in col_names):
